Fix parsing of BGP file in change_detector.py
There was a problem when reading the columns of a row.
The second column displaces the rest of the line with a newline
character at the end of the column when its width exceeds 17
characters. The new parsing method verifies the width of
the line read and if it is less than the width of the first
three rows it means the rest of the line was displaced, therefore
the next line must be read to complete the whole row.